docker-acts1/requirements.txt
requests
Django==2.2a1
djangorestframework
django-rest-auth